move the inline `app.rs` modules to `nu/`
