+++ src/lib.rs
@@ -3,6 +3,7 @@ mod app;
 mod config;
 mod edit;
 mod navigation;
+mod nu;
 mod terminal;
 mod tui;
